Changed name of python script, cleaned up code and changed README
